chat_bot/requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.0.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
lxml==4.9.2
pyTelegramBotAPI==4.9.0
python-telegram-bot==20.0
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
typing_extensions==4.4.0
urllib3==1.26.14
